requirements.txt
fastapi 
uvicorn 
requests 
transformers 
google-ads 
psycopg2